move condition.go to pkg/condition
